+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Updated to run with Minecraft 1.21 / Neoforge 21.0
 
+### Added
+
+- New consort dialogue
+- New carapacian dialogue
+
 ### Changed
 
 - Knitting needle crafting recipe now produces one item instead of two